modularization: extract summary generator
