+++ gcovr/summary_generator.py
@@ -1,0 +1,30 @@
+# -*- coding:utf-8 -*-
+
+# This file is part of gcovr <http://gcovr.com/>.
+#
+# Copyright 2013-2018 the gcovr authors
+# Copyright 2013 Sandia Corporation
+# This software is distributed under the BSD license.
+
+import sys
+
+from .utils import get_global_stats
+
+
+#
+# Prints a small report to the standard output
+#
+def print_summary(covdata):
+    (lines_total, lines_covered, percent,
+        branches_total, branches_covered,
+        percent_branches) = get_global_stats(covdata)
+
+    lines_out = "lines: %0.1f%% (%s out of %s)\n" % (
+        percent, lines_covered, lines_total
+    )
+    branches_out = "branches: %0.1f%% (%s out of %s)\n" % (
+        percent_branches, branches_covered, branches_total
+    )
+
+    sys.stdout.write(lines_out)
+    sys.stdout.write(branches_out)
